Split editor component
